% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_add_file_to_cache.R
\name{rb_add_file_to_cache}
\alias{rb_add_file_to_cache}
\title{Add file to cache}
\usage{
rb_add_file_to_cache(dribble, parent_id, project = NULL)
}
\arguments{
\item{dribble}{A dribble.}

\item{parent_id}{Identifier of the parent folder}

\item{project}{Defaults to NULL. Can be set once per session with
`rb_get_project_name()`. If given, must be a character vector of length
one: name of the project.}
}
\description{
Add file to cache
}
