% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_add_folder_to_cache.R
\name{rb_add_folder_to_cache}
\alias{rb_add_folder_to_cache}
\title{Add folder to cache}
\usage{
rb_add_folder_to_cache(
  dribble,
  parent_id,
  relative_path,
  project = NULL,
  cache = TRUE
)
}
\arguments{
\item{dribble}{A dribble.}

\item{parent_id}{Identifier of the parent folder.}

\item{relative_path}{Path relative to base folder.}

\item{project}{Defaults to NULL. Can be set once per session with
`rb_get_project_name()`. If given, must be a character vector of length
one: name of the project.}
}
\description{
Add folder to cache
}
