% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_drive_backup.R
\name{rb_backup}
\alias{rb_backup}
\title{Backup files}
\usage{
rb_backup(
  path,
  project,
  first_level_folders = NULL,
  glob = NULL,
  recurse = TRUE,
  create = TRUE,
  cache = TRUE,
  base_folder = "rbackupr"
)
}
\arguments{
\item{path}{Local path where files to backup are stored.}

\item{project}{Name of backup project}

\item{first_level_folders}{Defaults to NULL. If included, clarifies which folders within the path should be uploaded, keeping the folder structure.}

\item{glob}{Defaults to NULL. Can be used to filter type of files to upload, e.g. "*.jpg"}

\item{recurse}{Defaults to TRUE. Recurse up to one level.}

\item{create}{Logical, defaults to TRUE. Create folders if missing.}

\item{cache}{Logical, defaults to TRUE. Stores locally cached information about base and project folder.}

\item{base_folder}{Name of base folder, defaults to `rbackupr`}
}
\value{

}
\description{
Backup files
}
\examples{
\dontrun{
if (interactive()) {
rb_backup()
}
}
}
