% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_caching.R
\name{rb_check_cache_folder}
\alias{rb_check_cache_folder}
\title{Checks if cache folder exists, if not returns an informative message}
\usage{
rb_check_cache_folder()
}
\value{
If the cache folder exists, returns TRUE. Otherwise throws an error.
}
\description{
Checks if cache folder exists, if not returns an informative message
}
\examples{

# If cache folder does not exist, it throws an error
tryCatch(rb_check_cache_folder(),
  error = function(e) {
    return(e)
  }
)

# Create cache folder
rb_set_cache_folder(path = fs::path(
  tempdir(),
  "rb_cache_folder"
))
rb_create_cache_folder(ask = FALSE)

rb_check_cache_folder()
}
