% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_caching.R
\name{rb_create_cache_folder}
\alias{rb_create_cache_folder}
\title{Creates the base cache folder where `rbackupr` caches data.}
\usage{
rb_create_cache_folder(ask = TRUE)
}
\arguments{
\item{ask}{Logical, defaults to TRUE. If FALSE, and cache folder does not
exist, it just creates it without asking (useful for non-interactive
sessions).}
}
\value{
Nothing, used for its side effects.
}
\description{
Creates the base cache folder where `rbackupr` caches data.
}
\examples{
\donttest{
if (interactive()) {
  rb_create_cache_folder()
}
}
}
