% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_caching.R
\name{rb_disable_cache}
\alias{rb_disable_cache}
\title{Disable caching for the current session}
\usage{
rb_disable_cache()
}
\value{
Nothing, used for its side effects.
}
\description{
Disable caching for the current session
}
\examples{
\donttest{
if (interactive()) {
  rb_disable_cache()
}
}
}
