% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_drive_create_folders.R
\name{rb_drive_create_folders}
\alias{rb_drive_create_folders}
\title{Create missing folders and get data frame with identifiers of all given folders}
\usage{
rb_drive_create_folders(folders, parent_id, project = NULL, update = FALSE)
}
\arguments{
\item{folders}{A character vector of folder names.}

\item{parent_id}{Identifier of parent folder, where the new folders are to be created.}

\item{project}{Defaults to NULL. Can be set once per session with
`rb_get_project_name()`. If given, must be a character vector of length
one: name of the project.}

\item{update}{Logical, defaults to FALSE. If TRUE, checks on Google Drive for
newly updated files or folders, otherwise it assumes that only files and
folders listed in cache exist online.}
}
\value{

}
\description{
Create missing folders and get data frame with identifiers of all given folders
}
\examples{

if (interactive()) {

rb_drive_create_folders(folders = c("folder_a", "folder_b"),
                        parent_id = rb_get_project())
}

}
