% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_drive_backup.R
\name{rb_drive_find_base_folder}
\alias{rb_drive_find_base_folder}
\title{Create or get base folder}
\usage{
rb_drive_find_base_folder(base_folder = "rbackupr", cache = TRUE)
}
\arguments{
\item{base_folder}{Name of base folder, defaults to "rbackupr".}

\item{cache}{Logical, defaults to TRUE.}
}
\value{

}
\description{
Create or get base folder
}
\examples{
\dontrun{
rb_drive_find_base_folder()
}

}
