% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_drive_project.R
\name{rb_drive_find_project}
\alias{rb_drive_find_project}
\alias{rb_drive_create_project}
\title{Create rbackupr project}
\usage{
rb_drive_find_project(
  project = NULL,
  base_folder = "rbackupr",
  create = TRUE,
  cache = NULL
)

rb_drive_create_project(
  project = NULL,
  base_folder = "rbackupr",
  create = TRUE,
  cache = NULL
)
}
\arguments{
\item{project}{Defaults to NULL. Can be set once per session with `rb_get_project()`. If given, must be a character vector of length one: name of the project.}

\item{create}{Logical, defaults to TRUE. Creates folder if not existing.}
}
\value{
A dribble corresponding to the project folder.
}
\description{
Create rbackupr project
}
\examples{
\dontrun{
if (interactive()) {
  rb_drive_find_project(project = "example")
 }
}

rb_drive_create_project(project = "example")
}
