% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_caching.R
\name{rb_enable_cache}
\alias{rb_enable_cache}
\title{Enable caching for the current session}
\usage{
rb_enable_cache()
}
\value{
Nothing, used for its side effects.
}
\description{
Enable caching for the current session
}
\examples{
rb_enable_cache()
}
