% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_caching.R
\name{rb_get_cache_file}
\alias{rb_get_cache_file}
\title{Gets location of cache file}
\usage{
rb_get_cache_file(
  filename = "rbackupr_cache.sqlite",
  cache_folder = rbackupr::rb_get_cache_folder()
)
}
\arguments{
\item{filename}{Defaults to "rbackupr_cache.sqlite".}

\item{cache_folder}{Defaults to folder set with `rb_set_cache_folder()`}
}
\value{
A character vector of length one with location of item cache file.
}
\description{
Gets location of cache file
}
\examples{

rb_set_cache_folder(path = tempdir())
sqlite_cache_file_location <- rb_get_cache_file() # outputs location of cache file
sqlite_cache_file_location
}
