% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_caching.R
\name{rb_get_cache_table_name}
\alias{rb_get_cache_table_name}
\title{Gets name of table inside the database}
\usage{
rb_get_cache_table_name(type = "project", project = rbackupr::rb_get_project())
}
\arguments{
\item{type}{Defaults to "project". Type of cache file to output. Values
typically used by `rbackupr` include "base_folder", "projects", and
"project".}

\item{project}{Defaults to project name set with
`rbackupr::rb_get_project()`. Ignored if the parameter type is not set to
"project"}
}
\value{
A character vector of length one with the name of the relevant table
  in the cache file.
}
\description{
Gets name of table inside the database
}
\examples{
# outputs name of table used in the cache database
rb_get_cache_table_name(type = "project", language = "testing_project")
}
