% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_update_folder_cache.R
\name{rb_get_file_cache}
\alias{rb_get_file_cache}
\title{Check if new files appeared inside an online folder}
\usage{
rb_get_file_cache(dribble_id, update = FALSE, project = NULL, cache = TRUE)
}
\arguments{
\item{dribble_id}{The dribble identifier of a folder on Google Drive.}

\item{update}{Logical, defaults to FALSE. If TRUE, checks on Google Drive for newly updated folders.}

\item{project}{Defaults to NULL. Can be set once per session with
`rb_get_project()`. If given, must be a character vector of length one:
name of the project.}

\item{cache}{Logical, defaults to TRUE. Stores locally cached information
about base and project folder.}
}
\value{
A data frame with three columns: name (of the folder), id (of the
  dribble of the folder), parent_id (dribble id of the parent folder)
}
\description{
Check if new files appeared inside an online folder
}
\examples{

if (interactive()) {
  rb_get_file_cache(rb_drive_find_project())
}
}
