% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_get_cache.R
\name{rb_get_parent_folder}
\alias{rb_get_parent_folder}
\title{Gets parent id of a given folder}
\usage{
rb_get_parent_folder(dribble_id, project = NULL, base_folder = "rbackupr")
}
\arguments{
\item{dribble_id}{The dribble identifier of a folder on Google Drive.}

\item{project}{Defaults to NULL. Can be set once per session with
`rb_get_project_name()`. If given, must be a character vector of length
one: name of the project.}

\item{cache}{Logical, defaults to TRUE. Stores locally cached information
about base and project folder.}
}
\value{
A data frame with three columns: name (of the folder), id (of the
  dribble of the folder), parent_id (dribble id of the parent folder)
}
\description{
Gets parent id of a given folder
}
\examples{

if (interactive()) {
  rb_get_parent_folder(rb_drive_find_project())
}
}
