% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_get_cache.R
\name{rb_get_relative_path}
\alias{rb_get_relative_path}
\title{Get relative path in local folder from identifier}
\usage{
rb_get_relative_path(
  dribble_id,
  project = NULL,
  max_level = 100,
  base_folder = "rbackupr"
)
}
\arguments{
\item{dribble_id}{The dribble identifier of a folder on Google Drive.}

\item{project}{Defaults to NULL. Can be set once per session with
`rb_get_project_name()`. If given, must be a character vector of length
one: name of the project.}

\item{max_level}{Defaults to 100. Maximum level of sub-folders to backup.
Default means it will go 100 times deep into sub-folders. Used also to
prevent infinite loops.}

\item{base_folder}{Name of base folder, defaults to `rbackupr`}
}
\description{
Based only on cached data-
}
