% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_caching.R
\name{rb_set_project}
\alias{rb_set_project}
\alias{rb_get_project}
\title{Set (or get) name of project for the current session.}
\usage{
rb_set_project(project = NULL)

rb_get_project(project = NULL)
}
\arguments{
\item{project}{Defaults to NULL. If given, it must be a character vector of
length one. Name of a project. It will be used as the root folder for your
current project, and located under the `base_folder` on your Google Drive.}
}
\value{
The project name, if previously set; the same as input if not NULL;
  or the default, `rbackupr_data` is none is given.
}
\description{
Set (or get) name of project for the current session.
}
\examples{
rb_set_project(project = "weather_csv_files")
rb_get_project()
}
