% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_drive_auth.R
\name{rb_drive_auth}
\alias{rb_drive_auth}
\title{Set up app and scope for the current session}
\usage{
rb_drive_auth(
  app = NULL,
  scopes = "https://www.googleapis.com/auth/drive.file"
)
}
\arguments{
\item{app}{A Google app. See example, and dedicated gargle documentation.}

\item{scopes}{Defaults to `https://www.googleapis.com/auth/drive.file`. See Google api documentation for details}
}
\value{
Nothing, used for its side effects, i.e. logging in your Google account.
}
\description{
Set up app and scope for the current session
}
\examples{
if (interactive()) {
google_app <- httr::oauth_app(
"my-very-own-google-app",
key = "123456789.apps.googleusercontent.com",
secret = "abcdefghijklmnopqrstuvwxyz"
)
rb_drive_auth(app = google_app)
}
}
