% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_drive_project.R
\name{rb_drive_find_base_folder}
\alias{rb_drive_find_base_folder}
\title{Create or get base folder}
\usage{
rb_drive_find_base_folder(
  base_folder = "rbackupr",
  create = FALSE,
  cache = TRUE
)
}
\arguments{
\item{base_folder}{Name of base folder, defaults to "rbackupr".}

\item{create}{Defaults to FALSE. If set to TRUE, the folder is create if not found.}

\item{cache}{Logical, defaults to TRUE. Can be se to NULL, and managed with `rb_enable_cache()` / `rb_disable_cache()`}
}
\value{

}
\description{
Create or get base folder
}
\examples{
\dontrun{
if (interactive()) {
  rb_drive_find_base_folder()
}
}
}
